provide initial languages.lst and do not overwrite unless unless the list changes

+++ kamikaze/package/webif/files/usr/lib/webif/webif-mkthemelist.sh
@@ -2,6 +2,10 @@
 themes_lst="/etc/themes.lst"
 tmpthemeslst=$(mktemp "/tmp/.webif-XXXXXX")
 ipkg list | awk '/webif-theme/ { gsub("webif-theme-",""); print "option|"$1"|"$5}' | sort | uniq >> "$tmpthemeslst"
-rm -f "$themes_lst"
-mv -f "$tmpthemeslst" "$themes_lst"
-.
+if [ "'cat $tmpthemeslst'" != "" ]; then
+	if [ "'cat $tmpthemeslst'" != "'cat $themes_lst'" ]; then
+		rm -f "$themes_lst"
+		mv -f "$tmpthemeslst" "$themes_lst"
+	fi
+fi
+rm -f $tmpthemeslst
